haman_porfolio/portfolio/backend/requirements_prod.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
PyJWT==2.10.1
psycopg2-binary==2.9.11
marshmallow==3.20.1
marshmallow-sqlalchemy==1.4.2